+++ lib/abacate_pay/clients/payment_link_client.rb
@@ -23,7 +23,7 @@ module AbacatePay
       # @return [Array<Resources::Checkouts>]
       def list(**params)
         response = request("GET", "list", params: params.empty? ? nil : params)
-        Array(response).map { |data| Resources::Checkouts.new(data) }
+        build_list(response, Resources::Checkouts)
       end
 
       # @param id [String] The payment link ID